requirements.txt
streamlit==1.29.0
python-docx==0.8.11
langchain==0.1.0
langchain-openai==0.0.2
openai==1.7.0
